+++ src/components/index.js
@@ -13,3 +13,4 @@ export { default as ChatWidget } from './ChatWidget';
 export { default as ContactForm } from './ContactForm';
 export { default as ContactFields } from './ContactFields';
 export { default as ThankYou } from './ThankYou';
+export { default as PageIndicator } from './PageIndicator';